Add compatability with seblod CCK
There is a 'conflict' between readmore and seblod's cck.  seblod inserts some info in the introtext which allow it to link its own content.  These appear as ::cck::99::/cck:: and then the intro and full text is normally moved into full text and split using ::introtext::Some text::/introtext:: and ::fulltext::More text::/fulltext::

The changes will check if seblod is installed and if it is will remove the cck line and then strip off the introtext and fulltext tags.